distributed-task-queue/requirements.txt
fastapi==0.115.2
uvicorn==0.31.1
kafka-python-ng==2.2.3
pydantic==2.9.2
requests==2.32.3
